requirements.txt
asgiref==3.8.1
Django==4.2.14
django-cors-headers==4.4.0
django-dirtyfields==1.9.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
setuptools==72.1.0
sqlparse==0.5.1
uritemplate==4.1.1
